Updated test features
